Ata-391
Creating a new entity Student:
Created a new Student entity representing a student.

Adding a Data Access Object (DAO) for Student:
Created a DAO for the Student entity to provide access to the database

Creating a new StudentService and implementing it:
Created a StudentService that serves as an intermediate between controllers (or other components) and the DAO. The service implements business logic and operations on students using DAO methods.

Integration of the new command into TelegramBot:
Added code to the TelegramBot class that allows the bot to respond to the new command - ДобавитьУченика  This code will process incoming requests, extract the necessary information about the new student and pass it to the StudentService.

Testing and validation:
Conducted testing to make sure the bot responds to the command, data is written correctly through StudentService and saved using DAO.

New token for the bot:
The new token for the bot is 6386064861:AAGF_Gpdz-2WFFYC7zO9Sjdy_1HHifhzJtM